mediahub-backend/requirements.txt
uvicorn
path
fastapi
python-dotenv
pydantic
pydantic-settings